Refactor login
