+++ src/pages/authentication/index.ts
@@ -1,4 +1,4 @@
-export {default as LoginForm} from './LoginForm';
+export {default as LoginForm} from './login/login';
 export {default as RegisterForm} from './RegisterForm';
 export {default as RequestPasswordForm} from './RequestPasswordForm';
 export {default as SetPasswordForm} from './SetPasswordForm';
